# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.ExerciseRequest:
    properties:
      category_id:
        type: integer
      description:
        type: string
      name:
        type: string
    type: object
  models.ExerciseResponse:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://www.example.com/support
  description: This is a API for managing online workout tracker
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Online Workout Tracker API
  version: "1.0"
paths:
...
  /exercises:
    get:
      consumes:
      - application/json
      description: Get all exercises from the database
      produces:
      - application/json
      responses:
        "200":
          description: List of exercises
          schema:
            items:
              $ref: '#/definitions/models.ExerciseResponse'
            type: array
        "400":
          description: Request cancelled
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Exercises not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "504":
          description: Request timeout
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get all exercises
      tags:
      - exercises
    post:
      consumes:
      - application/json
      description: Create new exercise
      parameters:
      - description: Exercise data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.ExerciseRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Exercise created
          schema:
            $ref: '#/definitions/models.ExerciseResponse'
        "400":
          description: Request cancelled
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Exercise already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Failed to save exercise
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "504":
          description: Request timeout
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create exercise
      tags:
      - exercises
  /exercises/{id}:
    delete:
      consumes:
      - application/json
      description: Delete exercise by id
      parameters:
      - description: Exercise id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Exercise successfully deleted
        "400":
          description: Request cancelled
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Exercise not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Failed to delete exercise
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "504":
          description: Request timeout
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete exercise
      tags:
      - exercises
    get:
      consumes:
      - application/json
      description: Get exercise by id
      parameters:
      - description: Exercise id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Exercise data
          schema:
            $ref: '#/definitions/models.ExerciseResponse'
        "400":
          description: Request cancelled
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Exercise not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Error data receiving
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "504":
          description: Request timeout
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get exercise
      tags:
      - exercises
    put:
      consumes:
      - application/json
      description: Update exercise by id
      parameters:
      - description: Exercise id
        in: path
        name: id
        required: true
        type: integer
      - description: Exercise data
        in: body
        name: exercise
        required: true
        schema:
          $ref: '#/definitions/models.ExerciseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Exercise successfully updated
          schema:
            $ref: '#/definitions/models.ExerciseResponse'
        "400":
          description: Request cancelled
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Exercise not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: Exercise already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Failed to update exercise
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "504":
          description: Request timeout
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update exercise
      tags:
      - exercises
...
swagger: "2.0"